+++ CHANGELOG.md
@@ -5,7 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [0.8.0] - 2021-12-16
+- Fix ENS resolution to properly use reverse dns, simplify return object
+- Breaking change to ENS hook to only return the string name from the reverse ens domain.
 - Hide burned tokens from collection results in query
 
 ## [0.7.3] - 2021-12-08